[murmur/nginx] Add Mumble server setup
This adds a Mumble server with LDAP authentication. Data is currently
not persisted.

LDAP authentication is done with a script based on something in the
official Mumble repository, I made some changes to it to make it
slightly more sane.

As we are running on a Kubernetes trial we can currently only have one
static external IP address, so I can't set up a separate load balancer
for Mumble. To solve this for now I'm proxying it through nginx with a
connect timeout of 5 hours. Note that changing nginx config will
therefore currently cut Mumble connections. I'll add this to the docs.