[ejabberd/kaiwa] Move websockets to port 443
The idea behind this change is that users behind corporate firewalls may be able to connect through Kaiwa if the firewall can't tell Jabber websocket connections apart from other types of connections.

In addition the Kaiwa config is now in git and the session secret is stored in a configmap.